applications/python/requirements.txt
fabric-gateway==1.2.0
grpcio==1.60.0
protobuf==4.25.3
pyyaml==6.0.1
